move retirement to its own module
