+++ retirement/__init__.py
@@ -1,0 +1,1 @@
+from ._retirement import Retirement
