+++ user_input/__init__.py
@@ -1,0 +1,1 @@
+from ._user_input import InputHandler
