--- user_input_module/__init__.py
@@ -1,1 +1,0 @@
-from .user_input import InputHandler
